Modified: wflow-core - Check whether a directory is empty before delete it in FileUtil.
git-svn-id: http://dev.joget.org/svn/jw-community/trunk@406 1bff935d-2fe2-40b7-b00f-c2aca69a7612

+++ wflow-core/src/main/java/org/joget/apps/form/service/FileUtil.java
@@ -40,8 +40,11 @@ public class FileUtil implements ApplicationContextAware {
                     if (file != null) {
                         File tempFileDirectory = file.getParentFile();
                         FileUtil.storeFile(file, element, id);
-
-                        FileManager.deleteFile(tempFileDirectory);
+                        
+                        //check if directory is empty
+                        if (tempFileDirectory.listFiles().length == 0) {
+                            FileManager.deleteFile(tempFileDirectory);
+                        }
                     }
                 }
             }
